chore: move page components to page dir
