+++ packages/portal/tests/unit/components/page/PageHeader.spec.js
@@ -1,5 +1,5 @@
 import { createLocalVue, shallowMount } from '@vue/test-utils';
-import PageHeader from '@/components/PageHeader.vue';
+import PageHeader from '@/components/page/PageHeader.vue';
 import BootstrapVue from 'bootstrap-vue';
 
 const localVue = createLocalVue();
