+++ packages/portal/tests/unit/components/page/PageNavigation.spec.js
@@ -1,7 +1,7 @@
 import { createLocalVue, shallowMount } from '@vue/test-utils';
 
 import BootstrapVue from 'bootstrap-vue';
-import PageNavigation from '@/components/PageNavigation.vue';
+import PageNavigation from '@/components/page/PageNavigation.vue';
 
 const localVue = createLocalVue();
 localVue.use(BootstrapVue);
